backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.1
django-cache-url==3.4.5
django-cors-headers==4.4.0
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
environs==11.0.0
Markdown==3.7
marshmallow==3.22.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
